+++ tools/watcher/castor.php
@@ -9,7 +9,7 @@ use function Castor\parallel;
 use function Castor\run;
 
 #[AsTask(description: 'Build watcher for Linux system')]
-function linux()
+function linux(): void
 {
     run('go build -o bin/watcher-linux-amd64 -ldflags="-s -w" main.go', context: context()->withEnvironment(['GOOS' => 'linux', 'GOARCH' => 'amd64', 'CGO_ENABLED' => '0']));
     run('upx --brute bin/watcher-linux-amd64');
@@ -19,20 +19,20 @@ function linux()
 }
 
 #[AsTask(description: 'Build watcher for MacOS system')]
-function darwin()
+function darwin(): void
 {
     run('go build -o bin/watcher-darwin-amd64 -ldflags="-s -w" main.go', context: context()->withEnvironment(['GOOS' => 'darwin', 'GOARCH' => 'amd64', 'CGO_ENABLED' => '0']));
     run('go build -o bin/watcher-darwin-arm64 -ldflags="-s -w" main.go', context: context()->withEnvironment(['GOOS' => 'darwin', 'GOARCH' => 'arm64', 'CGO_ENABLED' => '0']));
 }
 
 #[AsTask(description: 'Build watcher for Windows system')]
-function windows()
+function windows(): void
 {
     run('go build -o bin/watcher-windows.exe -ldflags="-s -w" main.go', context: context()->withEnvironment(['GOOS' => 'windows', 'CGO_ENABLED' => '0']));
 }
 
 #[AsTask(description: 'Build watcher for all systems')]
-function build()
+function build(): void
 {
     parallel(static fn () => linux(), static fn () => darwin(), static fn () => windows());
 }